+++ seguros/documentos/templates/poliza/archivos_siniestros.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+{% extends 'base.html' %} 
+{% load crispy_forms_tags %} 
+{% block content %}
+
+        <h1>{{ titulo }}</h1>
+        <form method="post" enctype="multipart/form-data">
+            {% csrf_token %}
+            {% crispy form %}
+        </form>
+
+{% endblock %}
